requirements.txt
langchain
python-dotenv
langchain_community
beautifulsoup4
langchain_groq
langserve
sse_starlette
pandas
faiss-cpu
langchain_chroma
duckdb
duckduckgo-search
pymupdf
arxiv
wikipedia
